pyproject.toml
[project]
name = "multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.7",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.14",
    "langchain-text-splitters>=0.3.0",
    "langgraph>=1.1.6",
    "loguru>=0.7.3",
    "pydantic>=2.13.1",
    "python-dotenv>=1.0.0",
    "pytube>=15.0.0",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.56.0",
    "torch>=2.11.0",
    "youtube-transcript-api>=1.2.4",
]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "pytest>=8.0.0",
    "ruff>=0.11.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
